Use UniversalWrapperAspect as opposed to custom IdempotentAspect
